replaced homepage with new shadcn tables
